requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
openpyxl>=3.1.0
tencentcloud-sdk-python>=3.0.109
deep-translator>=1.11.4
